Restructured the handlers
