+++ services/py-api/src/server/routes/dependency_factory.py
@@ -32,7 +32,7 @@ def _h_service(
     return HackathonService(p_repo, t_repo, tx_manager)
 
 
-async def is_auth(authorization: Annotated[str, Header()]) -> None:
+def is_auth(authorization: Annotated[str, Header()]) -> None:
     # This follows the dependency pattern that is provided to us by FastAPI
     # You can read more about it here:
     # https://fastapi.tiangolo.com/tutorial/dependencies/dependencies-in-path-operation-decorators/
