+++ services/py-api/src/server/routes/teams_routes.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends
-from src.database.repository.teams_repository import TeamsRepository
-from src.server.handlers.teams_handlers import TeamHandlers
-from src.server.routes.dependency_factory import _t_repo
+from src.service.hackathon_service import HackathonService
+from src.server.handlers.hackathon_handlers import HackathonManagementHandlers
+from src.server.routes.dependency_factory import _h_service
 from src.server.schemas.response_schemas.schemas import ErrResponse, TeamDeletedResponse
 from starlette.responses import Response
 from src.server.routes.dependency_factory import is_auth, validate_obj_id
@@ -11,9 +11,9 @@ teams_router = APIRouter(prefix="/hackathon/teams")
 
 
 def _handler(
-    t_repo: TeamsRepository = Depends(_t_repo),
-) -> TeamHandlers:
-    return TeamHandlers(t_repo)
+    h_service: HackathonService = Depends(_h_service),
+) -> HackathonManagementHandlers:
+    return HackathonManagementHandlers(h_service)
 
 
 @teams_router.delete(
@@ -23,6 +23,6 @@ def _handler(
     dependencies=[Depends(is_auth), Depends(validate_obj_id)],
 )
 async def delete_team(
-    response: Response, object_id: str, handler: TeamHandlers = Depends(_handler)
+    response: Response, object_id: str, handler: HackathonManagementHandlers = Depends(_handler)
 ) -> TeamDeletedResponse | ErrResponse:
     return await handler.delete_team(response, object_id)
